Move js into lib folder
