+++ lib/PDFRStreamForBuffer.js
