+++ lib/PDFRStreamForFile.js
