+++ lib/PDFStreamForResponse.js
