+++ lib/PDFWStreamForBuffer.js
