+++ lib/PDFWStreamForFile.js
