+++ lib/muhammara.d.ts
