+++ lib/muhammara.js
@@ -2,7 +2,7 @@ var fs = require("fs");
 var path = require("path");
 var pregyp = require("@mapbox/node-pre-gyp");
 var binding_path = pregyp.find(
-  path.resolve(path.join(__dirname, "./package.json"))
+  path.resolve(path.join(__dirname, "../package.json"))
 );
 var muhammara = (module.exports = require(binding_path));
 var EventEmitter = require("events").EventEmitter;
