+++ tests/AppendSpecialPagesTest.js
@@ -2,7 +2,7 @@ var assert = require("assert");
 
 describe("AppendSpecialPagesTest", function () {
   it("should complete without error", function () {
-    var pdfWriter = require("../muhammara").createWriter(
+    var pdfWriter = require("../lib/muhammara").createWriter(
       __dirname + "/output/AppendSpecialPagesTest.pdf"
     );
