+++ tests/BasicJPGImagesTest.js
@@ -1,6 +1,6 @@
 describe("BasicJPGImagesTest", function () {
   it("should complete without error", function () {
-    var pdfWriter = require("../muhammara").createWriter(
+    var pdfWriter = require("../lib/muhammara").createWriter(
       __dirname + "/output/BasicJPGImagesTest.PDF"
     );
