+++ tests/DFontTest.js
@@ -1,6 +1,6 @@
 describe("DFontTest", function () {
   it("should complete without error", function () {
-    var pdfWriter = require("../muhammara").createWriter(
+    var pdfWriter = require("../lib/muhammara").createWriter(
       __dirname + "/output/DFontTest.pdf"
     );
     var page = pdfWriter.createPage(0, 0, 595, 842);
