+++ tests/EmptyPagesPDF.js
@@ -1,4 +1,4 @@
-var muhammara = require("../muhammara");
+var muhammara = require("../lib/muhammara");
 
 describe("EmptyPagesTest", function () {
   it("should complete without error", function () {
