+++ tests/PDFCopyingContextTest.js
@@ -1,6 +1,6 @@
 describe("PDFCopyingContextTest", function () {
   it("should complete without error", function () {
-    var pdfWriter = require("../muhammara").createWriter(
+    var pdfWriter = require("../lib/muhammara").createWriter(
       __dirname + "/output/PDFCopyingContextTest.PDF"
     );
     var copyingContext = pdfWriter.createPDFCopyingContext(
