+++ tests/SettingInfoValuesFromParsedContentTest.js
@@ -1,4 +1,4 @@
-var muhammara = require("../muhammara");
+var muhammara = require("../lib/muhammara");
 var assert = require("chai").assert;
 var fs = require("fs");
